swaptions/submission_swaptions/requirements.txt
numpy>=1.24,<2.1
pandas>=2.0,<2.3
scikit-learn>=1.3,<1.6
matplotlib>=3.7,<3.10
torch>=2.1,<2.5
merlinquantum>=0.1,<1.0
perceval-quandela>=0.10,<1.0
